Added info files and updated souce
Added Files: Authors, Contributors, Description. Fixed Readme a little. Deleted old fonts. Add some kerning to Source. Fixed Source to make export possible - but some glyph names that are not visually duplicates are being taged as duplicates in Glyphs. The source wont generate from Glyphs. This may not be an issue with fontmake. Will check with Joanna and Rainer and Mirko+Viviana